serverhop: Rewrite to use a SQL backend
The connect extension seems to have begun causing server hangs, so
switching to a push-pull database model is a more reliable solution,
with the caveat that the data is more likely to be out of date at any
given time.

This comes with one major caveat: There is no mechanism for resolving
domains, so the server hop config must only contain IP addresses. The
server likewise needs to be able to determine its own IP (via Steam or
status output) in order to update its own record in the database.